+++ x/evm/keeper/grpc_query.go
@@ -379,27 +379,29 @@ func (k Keeper) EstimateGasForEvmCallType(
 	// helper to check if a gas allowance results in an executable transaction.
 	executable := func(gas uint64) (vmError bool, rsp *evm.MsgEthereumTxResponse, err error) {
 		// update the message with the new gas value
-		evmMsg = gethcore.NewMessage(
-			evmMsg.From(),
-			evmMsg.To(),
-			evmMsg.Nonce(),
-			evmMsg.Value(),
-			gas,
-			evmMsg.GasPrice(),
-			evmMsg.GasFeeCap(),
-			evmMsg.GasTipCap(),
-			evmMsg.Data(),
-			evmMsg.AccessList(),
-			evmMsg.IsFake(),
-		)
+		evmMsg = core.Message{
+			To:                evmMsg.To,
+			From:              evmMsg.From,
+			Nonce:             evmMsg.Nonce,
+			Value:             evmMsg.Value,
+			GasLimit:          gas, // <---- This one changed
+			GasPrice:          evmMsg.GasPrice,
+			GasFeeCap:         evmMsg.GasFeeCap,
+			GasTipCap:         evmMsg.GasTipCap,
+			Data:              evmMsg.Data,
+			AccessList:        evmMsg.AccessList,
+			BlobGasFeeCap:     evmMsg.BlobGasFeeCap,
+			BlobHashes:        evmMsg.BlobHashes,
+			SkipAccountChecks: evmMsg.SkipAccountChecks,
+		}
 
 		tmpCtx := ctx
 		if fromType == evm.CallTypeRPC {
 			tmpCtx, _ = ctx.CacheContext()
 
-			acct := k.GetAccount(tmpCtx, evmMsg.From())
+			acct := k.GetAccount(tmpCtx, evmMsg.From)
 
-			from := evmMsg.From()
+			from := evmMsg.From
 			if acct == nil {
 				acc := k.accountKeeper.NewAccountWithAddress(tmpCtx, from[:])
 				k.accountKeeper.SetAccount(tmpCtx, acc)
@@ -412,7 +414,7 @@ func (k Keeper) EstimateGasForEvmCallType(
 				return true, nil, err
 			}
 			// resetting the gasMeter after increasing the sequence to have an accurate gas estimation on EVM extensions transactions
-			gasMeter := eth.NewInfiniteGasMeterWithLimit(evmMsg.Gas())
+			gasMeter := eth.NewInfiniteGasMeterWithLimit(evmMsg.GasLimit)
 			tmpCtx = tmpCtx.WithGasMeter(gasMeter).
 				WithKVGasConfig(storetypes.GasConfig{}).
 				WithTransientKVGasConfig(storetypes.GasConfig{})
@@ -495,7 +497,11 @@ func (k Keeper) TraceTx(
 		evmCfg.BaseFeeWei = baseFeeWeiPerGas
 	}
 
-	signer := gethcore.MakeSigner(evmCfg.ChainConfig, big.NewInt(ctx.BlockHeight()))
+	signer := gethcore.MakeSigner(
+		evmCfg.ChainConfig,
+		big.NewInt(ctx.BlockHeight()),
+		evm.ParseBlockTimeUnixU64(ctx),
+	)
 	txConfig := statedb.NewEmptyTxConfig(gethcommon.BytesToHash(ctx.HeaderHash().Bytes()))
 
 	// gas used at this point corresponds to GetProposerAddress & CalculateBaseFee
@@ -504,19 +510,19 @@ func (k Keeper) TraceTx(
 
 	for i, tx := range req.Predecessors {
 		ethTx := tx.AsTransaction()
-		msg, err := ethTx.AsMessage(signer, evmCfg.BaseFeeWei)
+		msg, err := core.TransactionToMessage(ethTx, signer, evmCfg.BaseFeeWei)
 		if err != nil {
 			continue
 		}
 		txConfig.TxHash = ethTx.Hash()
 		txConfig.TxIndex = uint(i)
 		// reset gas meter for each transaction
-		ctx = ctx.WithGasMeter(eth.NewInfiniteGasMeterWithLimit(msg.Gas())).
+		ctx = ctx.WithGasMeter(eth.NewInfiniteGasMeterWithLimit(msg.GasLimit)).
 			WithKVGasConfig(storetypes.GasConfig{}).
 			WithTransientKVGasConfig(storetypes.GasConfig{})
 		stateDB := statedb.New(ctx, &k, txConfig)
-		evmObj := k.NewEVM(ctx, msg, evmCfg, nil /*tracer*/, stateDB)
-		rsp, err := k.ApplyEvmMsg(ctx, msg, evmObj, nil /*tracer*/, false /*commit*/, txConfig.TxHash)
+		evmObj := k.NewEVM(ctx, *msg, evmCfg, nil /*tracer*/, stateDB)
+		rsp, err := k.ApplyEvmMsg(ctx, *msg, evmObj, nil /*tracer*/, false /*commit*/, txConfig.TxHash)
 		if err != nil {
 			continue
 		}
@@ -535,12 +541,12 @@ func (k Keeper) TraceTx(
 		tracerConfig, _ = json.Marshal(req.TraceConfig.TracerConfig)
 	}
 
-	msg, err := tx.AsMessage(signer, evmCfg.BaseFeeWei)
+	msg, err := core.TransactionToMessage(tx, signer, evmCfg.BaseFeeWei)
 	if err != nil {
 		return nil, err
 	}
 
-	result, _, err := k.TraceEthTxMsg(ctx, evmCfg, txConfig, msg, req.TraceConfig, tracerConfig)
+	result, _, err := k.TraceEthTxMsg(ctx, evmCfg, txConfig, *msg, req.TraceConfig, tracerConfig)
 	if err != nil {
 		// error will be returned with detail status from traceTx
 		return nil, err
@@ -605,19 +611,19 @@ func (k Keeper) TraceCall(
 	if err != nil {
 		return nil, grpcstatus.Errorf(grpccodes.Internal, "failed to unpack tx data: %s", err.Error())
 	}
-	evmMsg := gethcore.NewMessage(
-		gethcommon.HexToAddress(msgEthTx.From),
-		txData.GetTo(),
-		txData.GetNonce(),
-		txData.GetValueWei(),
-		txData.GetGas(),
-		txData.GetGasPrice(),
-		txData.GetGasFeeCapWei(),
-		txData.GetGasTipCapWei(),
-		txData.GetData(),
-		txData.GetAccessList(),
-		false, // isFake
-	)
+	evmMsg := core.Message{
+		To:                txData.GetTo(),
+		From:              gethcommon.HexToAddress(msgEthTx.From),
+		Nonce:             txData.GetNonce(),
+		Value:             txData.GetValueWei(), // amount
+		GasLimit:          txData.GetGas(),
+		GasPrice:          txData.GetGasPrice(),
+		GasFeeCap:         txData.GetGasFeeCapWei(),
+		GasTipCap:         txData.GetGasTipCapWei(),
+		Data:              txData.GetData(),
+		AccessList:        txData.GetAccessList(),
+		SkipAccountChecks: false,
+	}
 	result, _, err := k.TraceEthTxMsg(ctx, evmCfg, txConfig, evmMsg, req.TraceConfig, tracerConfig)
 	if err != nil {
 		// error will be returned with detail status from traceTx
@@ -679,7 +685,11 @@ func (k Keeper) TraceBlock(
 		tracerConfig, _ = json.Marshal(req.TraceConfig.TracerConfig)
 	}
 
-	signer := gethcore.MakeSigner(evmCfg.ChainConfig, big.NewInt(ctx.BlockHeight()))
+	signer := gethcore.MakeSigner(
+		evmCfg.ChainConfig,
+		big.NewInt(ctx.BlockHeight()),
+		evm.ParseBlockTimeUnixU64(ctx),
+	)
 	txsLength := len(req.Txs)
 	results := make([]*evm.TxTraceResult, 0, txsLength)
 
@@ -690,12 +700,12 @@ func (k Keeper) TraceBlock(
 		ethTx := tx.AsTransaction()
 		txConfig.TxHash = ethTx.Hash()
 		txConfig.TxIndex = uint(i)
-		msg, err := ethTx.AsMessage(signer, evmCfg.BaseFeeWei)
+		msg, err := core.TransactionToMessage(ethTx, signer, evmCfg.BaseFeeWei)
 		if err != nil {
 			result.Error = err.Error()
 			continue
 		}
-		traceResult, logIndex, err := k.TraceEthTxMsg(ctx, evmCfg, txConfig, msg, req.TraceConfig, tracerConfig)
+		traceResult, logIndex, err := k.TraceEthTxMsg(ctx, evmCfg, txConfig, *msg, req.TraceConfig, tracerConfig)
 		if err != nil {
 			result.Error = err.Error()
 		} else {
@@ -721,7 +731,7 @@ func (k *Keeper) TraceEthTxMsg(
 	ctx sdk.Context,
 	evmCfg statedb.EVMConfig,
 	txConfig statedb.TxConfig,
-	msg gethcore.Message,
+	msg core.Message,
 	traceConfig *evm.TraceConfig,
 	tracerJSONConfig json.RawMessage,
 ) (*any, uint, error) {
@@ -755,7 +765,9 @@ func (k *Keeper) TraceEthTxMsg(
 	}
 
 	if traceConfig.Tracer != "" {
-		if tracer, err = tracers.New(traceConfig.Tracer, tCtx, tracerJSONConfig); err != nil {
+		if tracer, err = tracers.DefaultDirectory.New(
+			traceConfig.Tracer, tCtx, tracerJSONConfig,
+		); err != nil {
 			return nil, 0, grpcstatus.Error(grpccodes.Internal, err.Error())
 		}
 	}
@@ -785,7 +797,7 @@ func (k *Keeper) TraceEthTxMsg(
 	// and not kvstore actions
 	// 3. Setup an empty transient KV gas config for transient gas to be
 	// calculated by opcodes
-	ctx = ctx.WithGasMeter(eth.NewInfiniteGasMeterWithLimit(msg.Gas())).
+	ctx = ctx.WithGasMeter(eth.NewInfiniteGasMeterWithLimit(msg.GasLimit)).
 		WithKVGasConfig(storetypes.GasConfig{}).
 		WithTransientKVGasConfig(storetypes.GasConfig{})
 	stateDB := statedb.New(ctx, k, txConfig)