+++ x/evm/keeper/vm_config.go
@@ -38,13 +38,7 @@ func (k *Keeper) TxConfig(
 func (k Keeper) VMConfig(
 	ctx sdk.Context, cfg *statedb.EVMConfig, tracer vm.EVMLogger,
 ) vm.Config {
-	var debug bool
-	if _, ok := tracer.(evm.NoOpTracer); !ok {
-		debug = true
-	}
-
 	return vm.Config{
-		Debug:     debug,
 		Tracer:    tracer,
 		NoBaseFee: false,
 		ExtraEips: cfg.Params.EIPs(),